requirements.txt
openai==0.27.4
python-docx==0.8.11
python-dotenv==1.0.0
python-telegram-bot==20.2
requests==2.28.2
Scrapy==2.8.0
schedule==1.1.0